Commiting Staged commits

+++ src/main/java/com/letswork/crm/entities/ConferenceRoom.java
@@ -1,15 +1,12 @@
 package com.letswork.crm.entities;
 
-import javax.persistence.Column;
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.validation.constraints.NotBlank;
-
-import com.letswork.crm.enums.ConferenceRoomType;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -36,7 +33,7 @@ public class ConferenceRoom extends Base{
 
     private String letsWorkCentre;
     
-    private Integer halfHourPrice;
+    private BigDecimal halfHourPrice;
     
 	private String state;
